Add option for scoped Cache locks
Before this change out of order releases of the same cache lock could lead
to situation where client A acquired the lock, took longer than the timeout,
which made client B successfully acquire the lock. If A now finishes while B
is still holding the lock A will release the lock that does not belong to it.

This fix introduces a unique value that is written as the cache value to stop
A from deleting B's lock.

+++ src/Illuminate/Cache/MemcachedLock.php
@@ -34,7 +34,7 @@ class MemcachedLock extends Lock
     public function acquire()
     {
         return $this->memcached->add(
-            $this->name, 1, $this->seconds
+            $this->name, $this->value(), $this->seconds
         );
     }
 
@@ -45,6 +45,18 @@ class MemcachedLock extends Lock
      */
     public function release()
     {
-        $this->memcached->delete($this->name);
+        if ($this->canRelease()) {
+            $this->memcached->delete($this->name);
+        }
+    }
+
+    /**
+     * Returns the value written into the driver for this lock.
+     *
+     * @return mixed
+     */
+    protected function getValue()
+    {
+        return $this->memcached->get($this->name);
     }
 }
